refactor configuration
