+++ src/views/dashboard/configuration/MetricOptions.vue
@@ -96,7 +96,7 @@ import {
   ChartTypes,
   PodsChartTypes,
   ListEntity,
-} from "../../data";
+} from "../data";
 import { ElMessage } from "element-plus";
 import Icon from "@/components/Icon.vue";
 import { useQueryProcessor, useSourceProcessor } from "@/hooks/useProcessor";
