+++ src/views/dashboard/configuration/StandardOptions.vue
@@ -117,7 +117,7 @@ limitations under the License. -->
 <script lang="ts" setup>
 import { ref } from "vue";
 import { useI18n } from "vue-i18n";
-import { SortOrder } from "../../data";
+import { SortOrder } from "../data";
 import { useDashboardStore } from "@/store/modules/dashboard";
 import { useQueryProcessor, useSourceProcessor } from "@/hooks/useProcessor";
 import { ElMessage } from "element-plus";
