+++ src/views/dashboard/configuration/WidgetOptions.vue
