+++ src/views/dashboard/configuration/graph-styles/Area.vue
