+++ src/views/dashboard/configuration/graph-styles/Bar.vue
