+++ src/views/dashboard/configuration/graph-styles/Card.vue
