+++ src/views/dashboard/configuration/graph-styles/EndpointList.vue
