+++ src/views/dashboard/configuration/graph-styles/HeatMap.vue
