+++ src/views/dashboard/configuration/graph-styles/InstanceList.vue
