+++ src/views/dashboard/configuration/graph-styles/Line.vue
