+++ src/views/dashboard/configuration/graph-styles/ServiceList.vue
