+++ src/views/dashboard/configuration/graph-styles/Table.vue
