+++ src/views/dashboard/configuration/graph-styles/TopList.vue
