+++ src/views/dashboard/configuration/graph-styles/index.ts
